refactor: taking out routes from his router folder
